@@ -1,6 +1,6 @@
 {
   "name": "react-native-material-ui",
-  "version": "1.12.0",
+  "version": "1.14.0",
   "description": "React Native Material Design Components",
   "main": "./index.js",
   "scripts": {
@@ -36,23 +36,25 @@
   "dependencies": {
     "color": "^1.0.3",
     "lodash.merge": "^4.0.0",
+    "prop-types": "^15.5.10",
     "react-native-material-design-styles": "^0.2.6"
   },
   "devDependencies": {
-    "babel-core": "^6.14.0",
-    "babel-eslint": "^7.1.1",
-    "babel-jest": "^17.0.2",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react-native": "^1.9.0",
-    "eslint": "^3.11.1",
-    "eslint-config-airbnb": "^13.0.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "^6.8.0",
-    "jest": "^17.0.3",
+    "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^20.0.3",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react-native": "^2.1.0",
+    "jest": "^20.0.4",
     "react": "16.0.0-alpha.6",
     "react-native": "^0.44.0",
-    "react-test-renderer": "^15.4.1"
+    "react-test-renderer": "^15.6.1",
+    "react-native-vector-icons": "^4.0.0",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.0.1",
+    "eslint-plugin-react": "^7.1.0"
   }
 }
