@@ -1,6 +1,11 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
-import { Platform, TouchableNativeFeedback, TouchableWithoutFeedback } from 'react-native';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
+import {
+    Platform,
+    TouchableNativeFeedback,
+    TouchableWithoutFeedback,
+} from 'react-native';
 /* eslint-enable import/no-unresolved, import/extensions */
 
 const propTypes = {
@@ -12,11 +17,12 @@ const propTypes = {
     children: PropTypes.node.isRequired,
 };
 const defaultProps = {
+    color: null,
     borderless: true,
 };
 
 function isCompatible() {
-    if (Platform.OS === 'ios') {
+    if (Platform.OS === 'ios' || Platform.OS === 'web') {
         return false;
     }
 
@@ -26,7 +32,6 @@ function isCompatible() {
 class RippleFeedback extends PureComponent {
     render() {
         const { children, color, borderless, ...otherProps } = this.props;
-
 
         if (!isCompatible()) {
             return (
@@ -50,7 +55,6 @@ class RippleFeedback extends PureComponent {
             </TouchableNativeFeedback>
         );
     }
-
 }
 
 RippleFeedback.propTypes = propTypes;
