@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unresolved, import/extensions */
-import React, { PureComponent, PropTypes } from 'react';
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
 import { View, StyleSheet, NativeModules, findNodeHandle } from 'react-native';
 /* eslint-enable import/no-unresolved, import/extensions */
 
@@ -20,7 +21,11 @@ const propTypes = {
     onSearchPress: PropTypes.func.isRequired,
 };
 const defaultProps = {
+    rightElement: null,
+    onRightElementPress: null,
+    size: null,
     style: {},
+    searchable: null,
 };
 const contextTypes = {
     uiTheme: PropTypes.object.isRequired,
@@ -109,7 +114,7 @@ class RightElement extends PureComponent {
 
                 return (
                     <IconToggle
-                        key={index}
+                        key={action}
                         name={action}
                         color={flattenRightElement.color}
                         size={size}
@@ -175,7 +180,7 @@ class RightElement extends PureComponent {
                         }}
                     />
                     <IconToggle
-                        name="more-vert"
+                        name={rightElement.menu.icon || 'more-vert'}
                         color={flattenRightElement.color}
                         size={size}
                         onPress={() => this.onMenuPressed(rightElement.menu.labels)}
